Working on Primary Color

+++ database/migrations/2023_12_13_194112_create_hostels_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('hostels', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('number');
+            $table->integer('number');
             $table->char('section')->nullable();
             $table->string('place')->default('محصولی');
             $table->integer('capacity')->default(5);
